模型端计算函数

+++ models/out_device.go
@@ -18,7 +18,7 @@ type OutdoorDevice struct {
 	Fo        float64 `json:"fo" gorm:"column:fo"`                 // 室外风机运转风速等级
 	TdSH      float64 `json:"tdsh" gorm:"column:tdsh"`             // 排气温度与饱和冷凝温度差值。TdSH = Td1-Tcond
 	Info1     float64 `json:"tsc" gorm:"column:tsc"`               // Te温度与饱和冷凝温度差值。TeSC =Tcond -Te
-	Status    int     `json:"ou_off" gorm:"column:ou_off"`         // 运行状态
+	Status    float64 `json:"ou_off" gorm:"column:ou_off"`         // 运行状态
 	Time      string  `json:"up_unix_time,omitempty" gorm:"-"`     // 时间字符串
 	TimeStamp int64   `json:"time" gorm:"column:up_unix_time"`     // 时间戳
 }
